Fixed PUND
PUND works now! The time array was just a bit messed up. Also figured out that the awg is fine if you give it max points at max frequency, it just does its darned best whatever you send it, what a trooper. This is good because that level of check probably should have been in the driver anyway.