Added virtual option to generic scpi drivers
Passing the address "VIRTUAL" will result in a virtual instance. This will pass rm writes and queries to a virtual rm class. The writes are just tossed, and the queries are just looked up from a dict. When the scope specifically is queried for its waveform an if statement is triggered and some virtual data saved locally on a csv is passed instead.